app/requirements.txt
streamlit
pandas
altair
google-cloud-bigquery
google-cloud-bigquery-storage
google-auth
pyarrow
db-dtypes